Consolidate the local and remote trackers
